requirements.txt
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
convertdate==2.4.0
et-xmlfile==1.1.0
flake8==7.1.1
idna==3.10
isort==5.13.2
logging==0.4.9.6
lunardate==0.2.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pyluach==2.2.0
PyMeeus==0.5.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
workalendar==17.0.0
